Fixed Bug: default image being deleted in profile edit when user has not uploaded a profile image while registering.

+++ profile-edit/includes/profile-edit.inc.php
@@ -116,12 +116,14 @@ if (isset($_POST['update-profile'])) {
                         *   Deleting old profile photo
                         * -------------------------------------------------------------------------------
                         */
-                        if (!unlink('../../assets/uploads/users/' . $_SESSION['profile_image'])) {  
+						if ( $_SESSION['profile_image'] != "_defaultUser.png" ) {
+							if (!unlink('../../assets/uploads/users/' . $_SESSION['profile_image'])) {  
 
-                            $_SESSION['ERRORS']['imageerror'] = 'old image could not be deleted';
-                            header("Location: ../");
-                            exit();
-                        } 
+								$_SESSION['ERRORS']['imageerror'] = 'old image could not be deleted';
+								header("Location: ../");
+								exit();
+							} 
+						}
                     }
                     else
                     {
